requirements.txt
einops
fairseq
flask
flask_cors
gin
gin_config
librosa
local_attention
matplotlib
numpy
praat-parselmouth
pyworld
PyYAML
resampy
scikit_learn
scipy
SoundFile
tensorboard
torchcrepe
tqdm
transformers
wave
pysimplegui
sounddevice
gradio
faiss-cpu
loguru
rich
PyYAML
